Better Typed and ReadOnly support
I made it easier to work with sublists and read-only lists.
I made it easier to work with sublists and typed lists.
I created the Typed extension method for creating TypedLists.
I handle the case where someone is reversed a concrete Sublist.
I made it easier to restored a Substring from an IReadOnlySublist.